auto_test/requirements.txt
selenium
pytest
pytest-rerunfailures
pytest-xdist
allure-pytest
pyyaml
requests